🚚 Rename compat.py to _compat.py
